python3 fix to compensate for change in integer division between py2, and py3

+++ cbm3_python/cbm3data/resultsloader.py
@@ -214,7 +214,7 @@ INSERT INTO tblDistIndicators
             accessDB.ExecuteQuery(qryFormatted)
 
     def printProgress(self, fileLen, fluxIndPath, recordNum):
-        if ((recordNum-1) % (fileLen/10)) == 0 or recordNum == fileLen:
+        if ((recordNum-1) % int(fileLen/10)) == 0 or recordNum == fileLen:
             logging.info("load file {0}: {1}/{2} ({3})".format(
                 fluxIndPath, recordNum, fileLen, 
                 self.floored_percentage(float(recordNum)/fileLen, 2)))
